[RESTEASY-1805] Extract interface from ResourceClass
