+++ resteasy-jaxrs/src/main/java/org/jboss/resteasy/spi/metadata/ResourceClass.java
@@ -1,62 +1,21 @@
 package org.jboss.resteasy.spi.metadata;
 
 /**
- * @author <a href="mailto:bill@burkecentral.com">Bill Burke</a>
- * @version $Revision: 1 $
+ * @author Christian Kaltepoth
  */
-public class ResourceClass
+public interface ResourceClass
 {
-   private static final FieldParameter[] EMPTY_FIELD_PARAMS = {};
-   private static final SetterParameter[] EMPTY_SETTER_PARAMETERS = {};
-   private static final ResourceMethod[] EMPTY_RESOURCE_METHODS = {};
-   private static final ResourceLocator[] EMPTY_RESOURCE_LOCATORS = {};
+  String getPath();
 
-   protected Class<?> clazz;
-   protected FieldParameter[] fields = EMPTY_FIELD_PARAMS;
-   protected SetterParameter[] setters = EMPTY_SETTER_PARAMETERS;
-   protected ResourceMethod[] resourceMethods = EMPTY_RESOURCE_METHODS;
-   protected ResourceLocator[] resourceLocators = EMPTY_RESOURCE_LOCATORS;
-   protected ResourceConstructor constructor; // only one allowed
-   protected String path;
+  Class<?> getClazz();
 
-   public ResourceClass(Class<?> clazz, String path)
-   {
-      this.clazz = clazz;
-      this.path = path;
-   }
+  ResourceConstructor getConstructor();
 
-   public String getPath()
-   {
-      return path;
-   }
+  FieldParameter[] getFields();
 
-   public Class<?> getClazz()
-   {
-      return clazz;
-   }
+  SetterParameter[] getSetters();
 
-   public ResourceConstructor getConstructor()
-   {
-      return constructor;
-   }
+  ResourceMethod[] getResourceMethods();
 
-   public FieldParameter[] getFields()
-   {
-      return fields;
-   }
-
-   public SetterParameter[] getSetters()
-   {
-      return setters;
-   }
-
-   public ResourceMethod[] getResourceMethods()
-   {
-      return resourceMethods;
-   }
-
-   public ResourceLocator[] getResourceLocators()
-   {
-      return resourceLocators;
-   }
+  ResourceLocator[] getResourceLocators();
 }
